Feature #177: change provides to binds
