+++ feature/main/src/main/java/com/hankki/feature/main/network/NetworkModule.kt
@@ -1,20 +1,18 @@
 package com.hankki.feature.main.network
 
-import android.content.Context
+import dagger.Binds
 import dagger.Module
-import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object NetworkModule {
+abstract class NetworkModule {
 
-    @Provides
+    @Binds
     @Singleton
-    fun provideNetworkMonitor(@ApplicationContext context: Context): NetworkMonitor {
-        return NetworkMonitor(context)
-    }
+    abstract fun bindNetworkMonitor(
+        networkMonitorImpl: NetworkMonitorImpl,
+    ): NetworkMonitor
 }
