+++ feature/main/src/main/java/com/hankki/feature/main/network/NetworkMonitor.kt
@@ -1,48 +1,8 @@
 package com.hankki.feature.main.network
 
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
-import android.net.NetworkRequest
-import dagger.hilt.android.qualifiers.ApplicationContext
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import javax.inject.Inject
+import kotlinx.coroutines.flow.StateFlow
 
-class NetworkMonitor @Inject constructor(
-    @ApplicationContext context: Context,
-) {
-    private val connectivityManager =
-        context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-
-    private val _isConnected = MutableStateFlow(false)
-    val isConnected = _isConnected.asStateFlow()
-
-    private val networkCallback = object : ConnectivityManager.NetworkCallback() {
-        override fun onAvailable(network: Network) {
-            _isConnected.value = true
-        }
-
-        override fun onLost(network: Network) {
-            _isConnected.value = false
-        }
-    }
-
-    init {
-        val networkRequest = NetworkRequest.Builder()
-            .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-            .build()
-
-        connectivityManager.registerNetworkCallback(networkRequest, networkCallback)
-
-        val activeNetwork = connectivityManager.activeNetwork
-        val capabilities = connectivityManager.getNetworkCapabilities(activeNetwork)
-        _isConnected.value =
-            capabilities?.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) == true
-    }
-
-    fun unregisterCallback() {
-        connectivityManager.unregisterNetworkCallback(networkCallback)
-    }
+interface NetworkMonitor {
+    val isConnected: StateFlow<Boolean>
+    fun unregisterCallback()
 }
